Added analytics support
The tracking key is saved in the config file. A rake task, 'analytics', is
created to enable or disable analytics.

In order to make analytics work, the build script should include the
`rake analytics` task before the `jekyll build` command. After the
command, the task should be executed again to disable analytics locally.