Split the build script from the publication on Travis
This allows running the publication only for branch pushes and not for
pull requests.
The creation of the orphan commit for the master branch with the built
website has been refactored to avoid having to filtering the branch each
time, which is faster.
The new script also avoids changing the git email and username globally,
making it much safer to run the publish task locally (through bldr for
instance).